requirements.txt
boilerpy3
streamlit
torch
transformers
bert-extractive-summarizer
